feat: improve runtime media and graphics fidelity

+++ refract/src/stubs/audio/sceAudioChangeChannelVolume.cpp
@@ -4,7 +4,8 @@ void sceAudioChangeChannelVolume(Implementation& implementation, psprecomp::Stat
 #if !defined(__PSP__)
   static_cast<void>(implementation);
   state.gpr[2] = audio_state::change_volume(
-      state.gpr[4], state.gpr[5], state.gpr[6]);
+      state.gpr[4], static_cast<std::int32_t>(state.gpr[5]),
+      static_cast<std::int32_t>(state.gpr[6]));
   return;
 #else
   (void)implementation;
