+++ refract/src/stubs/audio/sceAudioOutputPanned.cpp
@@ -3,8 +3,9 @@
 void sceAudioOutputPanned(Implementation& implementation, psprecomp::State& state) {
 #if !defined(__PSP__)
   static_cast<void>(implementation);
-  state.gpr[2] = audio_state::output(state, state.gpr[4], state.gpr[5],
-                                     state.gpr[6], state.gpr[7], false);
+  state.gpr[2] = audio_state::output(
+      state, state.gpr[4], static_cast<std::int32_t>(state.gpr[5]),
+      static_cast<std::int32_t>(state.gpr[6]), state.gpr[7], false);
   return;
 #else
   (void)implementation;
